simplified examples
